Added functionality to upload button, fixed some of the architecture and naming convention of code
