+++ src/model/Main.java
@@ -2,11 +2,9 @@ package model;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.image.Image;
 import javafx.stage.Stage;
-import view.MainController;
 
 public class Main extends Application {
