+++ src/view/HomePage.fxml
@@ -14,7 +14,7 @@
 <?import javafx.scene.shape.Rectangle?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane prefHeight="650.0" prefWidth="1200.0" style="-fx-background-color: #f7620e;" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="view.HomePageController">
+<AnchorPane prefHeight="650.0" prefWidth="1200.0" style="-fx-background-color: #f7620e;" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="view.HomePageLogic">
    <children>
       <MediaView fx:id="songPlayer" fitHeight="200.0" fitWidth="200.0" />
       <MediaView fx:id="homeVideo" fitHeight="600.0" fitWidth="1200.0" preserveRatio="false" />
